refactor: 이벤트 처리 메서드에서 진행하는 동작을 서비스로 이동
