+++ matzip-app-external-api/src/main/java/com/woowacourse/matzip/application/RestaurantEventListener.java
@@ -1,47 +1,42 @@
 package com.woowacourse.matzip.application;
 
-import com.woowacourse.matzip.domain.restaurant.RestaurantRepository;
 import com.woowacourse.matzip.domain.review.ReviewCreateEvent;
 import com.woowacourse.matzip.domain.review.ReviewDeleteEvent;
 import com.woowacourse.matzip.domain.review.ReviewUpdateEvent;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-@Service
+@Component
 public class RestaurantEventListener {
 
-    private final RestaurantRepository restaurantRepository;
+    private final RestaurantService restaurantService;
 
-    public RestaurantEventListener(final RestaurantRepository restaurantRepository) {
-        this.restaurantRepository = restaurantRepository;
+    public RestaurantEventListener(final RestaurantService restaurantService) {
+        this.restaurantService = restaurantService;
     }
 
     @Async
-    @Transactional
     @TransactionalEventListener
     public void handleReviewCreateEvent(final ReviewCreateEvent event) {
         Long restaurantId = event.getRestaurantId();
         int rating = event.getRating();
-        restaurantRepository.updateRestaurantByReviewInsert(restaurantId, rating);
+        restaurantService.updateWhenReviewCreate(restaurantId, rating);
     }
 
     @Async
-    @Transactional
     @TransactionalEventListener
     public void handleReviewDeleteEvent(final ReviewDeleteEvent event) {
         Long restaurantId = event.getRestaurantId();
         int rating = event.getRating();
-        restaurantRepository.updateRestaurantByReviewDelete(restaurantId, rating);
+        restaurantService.updateWhenReviewDelete(restaurantId, rating);
     }
 
     @Async
-    @Transactional
     @TransactionalEventListener
     public void handleReviewUpdateEvent(final ReviewUpdateEvent event) {
         Long restaurantId = event.getRestaurantId();
         int ratingGap = event.getRatingGap();
-        restaurantRepository.updateRestaurantByReviewUpdate(restaurantId, ratingGap);
+        restaurantService.updateWhenReviewUpdate(restaurantId, ratingGap);
     }
 }
