+++ matzip-app-external-api/src/test/java/com/woowacourse/matzip/application/RestaurantServiceTest.java
@@ -14,7 +14,10 @@ import com.woowacourse.matzip.domain.member.Member;
 import com.woowacourse.matzip.domain.member.MemberRepository;
 import com.woowacourse.matzip.domain.restaurant.Restaurant;
 import com.woowacourse.matzip.domain.restaurant.RestaurantRepository;
+import com.woowacourse.matzip.domain.review.ReviewCreateEvent;
+import com.woowacourse.matzip.domain.review.ReviewDeleteEvent;
 import com.woowacourse.matzip.domain.review.ReviewRepository;
+import com.woowacourse.matzip.domain.review.ReviewUpdateEvent;
 import com.woowacourse.matzip.exception.RestaurantNotFoundException;
 import com.woowacourse.support.SpringServiceTest;
 import java.util.List;
@@ -175,6 +178,87 @@ class RestaurantServiceTest {
                         .extracting("id")
                         .containsExactly(restaurant2.getId(), restaurant1.getId()),
                 () -> assertThat(response2.isHasNext()).isFalse()
+        );
+    }
+
+    @Test
+    void 리뷰_작성에_대해_음식점_정보를_수정한다() {
+        Restaurant restaurant = restaurantRepository.save(
+                Restaurant.builder()
+                        .campusId(1L)
+                        .categoryId(1L)
+                        .name("음식점")
+                        .address("주소")
+                        .kakaoMapUrl("URL")
+                        .imageUrl("이미지 주소")
+                        .distance(0)
+                        .build()
+        );
+
+        restaurantService.updateWhenReviewCreate(restaurant.getId(), 5);
+
+        Restaurant actual = restaurantRepository.findById(restaurant.getId())
+                .orElseThrow();
+        assertAll(
+                () -> assertThat(actual.getReviewCount()).isOne(),
+                () -> assertThat(actual.getReviewRatingSum()).isEqualTo(5),
+                () -> assertThat(actual.getReviewRatingAverage()).isEqualTo(5.0f)
+        );
+    }
+
+    @Test
+    void 리뷰_삭제_이벤트를_바탕으로_음식점_정보를_수정한다() {
+        Restaurant restaurant = restaurantRepository.save(
+                Restaurant.builder()
+                        .campusId(1L)
+                        .categoryId(1L)
+                        .name("음식점")
+                        .address("주소")
+                        .kakaoMapUrl("URL")
+                        .imageUrl("이미지 주소")
+                        .distance(0)
+                        .reviewCount(1)
+                        .reviewRatingSum(5)
+                        .reviewRatingAverage(5.0f)
+                        .build()
+        );
+
+        restaurantService.updateWhenReviewDelete(restaurant.getId(), 5);
+
+        Restaurant actual = restaurantRepository.findById(restaurant.getId())
+                .orElseThrow();
+        assertAll(
+                () -> assertThat(actual.getReviewCount()).isZero(),
+                () -> assertThat(actual.getReviewRatingSum()).isZero(),
+                () -> assertThat(actual.getReviewRatingAverage()).isZero()
+        );
+    }
+
+    @Test
+    void 리뷰_수정_이벤트를_바탕으로_음식점_정보를_수정한다() {
+        Restaurant restaurant = restaurantRepository.save(
+                Restaurant.builder()
+                        .campusId(1L)
+                        .categoryId(1L)
+                        .name("음식점")
+                        .address("주소")
+                        .kakaoMapUrl("URL")
+                        .imageUrl("이미지 주소")
+                        .distance(0)
+                        .reviewCount(1)
+                        .reviewRatingSum(5)
+                        .reviewRatingAverage(5.0f)
+                        .build()
+        );
+
+        restaurantService.updateWhenReviewUpdate(restaurant.getId(), -3);
+
+        Restaurant actual = restaurantRepository.findById(restaurant.getId())
+                .orElseThrow();
+        assertAll(
+                () -> assertThat(actual.getReviewCount()).isOne(),
+                () -> assertThat(actual.getReviewRatingSum()).isEqualTo(2),
+                () -> assertThat(actual.getReviewRatingAverage()).isEqualTo(2.0f)
         );
     }
 }
